requirements.txt
Django
wagtail>=6.0
django-allauth
djangorestframework